slack_aws_snowflake_pipeline/lambda_module/resources/requirements.txt
slack-sdk
requests
pandas
python-dotenv